Updated python files format using Pylint
