+++ .github/workflows/update-project-stats.yml
@@ -28,4 +28,4 @@ jobs:
           pip3 install pyyaml
           pip3 install requests
           pip3 install beautifulsoup4
-          python fork_count/updateForkCount.py
+          python fork_count/update_fork_count.py
